fix: 비교 함수 모듈화
